Store::SRM: fix reading with blocks without chunks
Starting with 6.41.04 SRMwin inserts a block with no chunks and the
timestamp of the last block. As this wasn't expected, the code did read a
chunk, for this block, anyways. This lead to mismatching timestamps and
aborted file reading.

Fixed the code to cope with zero-length blocks - although this looks a bit
broken.