Store::SRM: support SRM5 file format
SRM5 is basically the same as SRM6, but without "blocks" and with a marker
comment length of 3 chars. Due to the lack of blocks, SRM5 only specifies
the date of the recording - the actual time isn't recoreded. I'm also
guessing, that due to the lack of blocks, recording pauses aren't
identifiable, as well.