Store::SRM: ignore trailing data for SRM7 files
SRMWin 6.42.08 writes files with some unrecognized bytes at the end.
We used to keep reading to recover files with inconsistent chunk count.
Unfortunatly this is now causing bad data at the end.

So, we're now sticking to the given chunk count for SRM7 files as we do
for SRM4/5/6, already.