Fit: handle broken start events more graceful
All records with a timestamp predating a start event were dropped. This
showed up with a broken file having a start event in the ultimate future.

We're now handling this more graceful by just dropping the first record
after a bad start event. We can't avoid dropping the first record as this
is part of the gap detection.